% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependency_helper.R
\name{append_grid_}
\alias{append_grid_}
\title{Internal function for use by generate_grid family of functions}
\usage{
append_grid_(feature_dt, feature_col, grid)
}
\arguments{
\item{feature_dt}{data.table containing features used in predictive model}

\item{feature_col}{character. name of a column in \code{feature_dt}}
}
\description{
Replaces \code{feature_col} with \code{grid} values and maintains column order
}
\examples{
\dontrun{
dt <- data.table(a = 1:3, b = 4:6)
grid <- 10:12
append_grid_(dt, "a", grid)
}
}

