% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variable_importance.R
\name{base_model_loss_}
\alias{base_model_loss_}
\title{Internal function for use by calculate_marginal_vimp}
\usage{
base_model_loss_(method, x, y, resampling_indices, trControl, tuneGrid,
  loss_metric, ...)
}
\arguments{
\item{method}{character string defining method to pass to caret}

\item{x}{data.table containing predictor variables}

\item{y}{vector containing target variable}

\item{resampling_indices}{a list of integer vectors corresponding to the row indices
used for each resampling iteration}

\item{trControl}{trainControl object to be passed to caret train.}

\item{tuneGrid}{a data.frame containing hyperparameter values for caret.
Should only contain one value for each hyperparameter. Set to NULL
if caret method does not have any hyperparameter values.}

\item{loss_metric}{character. Loss metric to evaluate accuracy of model}

\item{...}{additional arguments to pass to caret train}
}
\description{
Calculate baseline accuracy using all variables in training data
}

