% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variable_importance.R
\name{calculate_marginal_vimp}
\alias{calculate_marginal_vimp}
\title{Calculate the importance of all predictor variables in a training set}
\usage{
calculate_marginal_vimp(x, y, method, loss_fcn, resampling_indices, tuneGrid,
  vars = names(x), seed = sample(.Random.seed, 1), ...)
}
\arguments{
\item{x}{data.table containing predictor variables}

\item{y}{vector containing target variable}

\item{method}{character string defining method to pass to caret}

\item{loss_fcn}{loss function to evaluate accuracy of model}

\item{resampling_indices}{a list of integer vectors corresponding to the rows
used for each training iteration}

\item{tuneGrid}{a data.frame containing hyperparameter values for caret.
Should only contain one value for each hyperparamter}

\item{vars}{character vector specifying variables for which to determine importance
Defaults to all predictor variables in \code{x}.}

\item{seed}{set random seed for train method}

\item{...}{additional arguments to pass to caret train}
}
\description{
Deterimne variable importance of predictor variables by observing
the change in error when removing a single variable at a time
from training data
}

