% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependency_individual.R
\name{calculate_pd_vimp}
\alias{calculate_pd_vimp}
\title{Calculate variable importance based on partial dependency}
\usage{
calculate_pd_vimp(pd, vimp_colname = "ensemble")
}
\arguments{
\item{pd}{output from \code{\link{calculate_partial_dependency}}}

\item{vimp_colname}{name of model (taken from the column names in \code{pd})
for which to calculate variable importance}
}
\description{
Variable importance is calculated as the difference between the min and max
of the partial dependency output for a given feature
}
\examples{
\dontrun{
# Example output from calculate_partial_dependency
pd <- data.table(feature = rep("a", 9),
                 feature_val = rep(c(1, 3.5, 6), 3),
                 model = rep(c("model1",
                               "model2",
                               "ensemble"),
                             each = 3),
                 prediction = c(c(-2.5, 0, 2.5),
                                c(0, 0, 0),
                                c(-2.5, -0.75, 0)))
calculate_pd_vimp(pd, vimp_colname = "ensemble")
}
}

