% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependency_aggregate.R
\name{facet_plot_fcn}
\alias{facet_plot_fcn}
\title{Multiple partial dependency plots on one page}
\usage{
facet_plot_fcn(pd_list, plot_fcn = plot_partial_dependency, nrow = NULL,
  ncol = NULL, scales = "free")
}
\arguments{
\item{pd_list}{output from \code{\link{loop_calculate_partial_dependency}}}

\item{plot_fcn}{a function that accepts the output from
\code{\link{calculate_partial_dependency}} and returns a ggplot object.
This function will be looped over \code{pd_list}.}

\item{nrow}{int. Number of rows in layout}

\item{ncol}{int. Number of columns in layout}

\item{scales}{one of "\code{fixed}", "\code{free}", "\code{free_x}", "\code{free_y}".
See \code{ggplot2::facet_wrap} for more information}
}
\description{
It can be helpful to view multiple partial dependency plots at once
to get a sense of variable importance, relationships between variables, etc.
}

