% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependency_helper.R
\name{generate_grid_range_categorical_}
\alias{generate_grid_range_categorical_}
\title{Internal function for use by calculate_partial_dependency}
\usage{
generate_grid_range_categorical_(feature_dt, feature_col, custom_range = NULL)
}
\arguments{
\item{feature_dt}{data.table containing features used in predictive model}

\item{feature_col}{character. name of a column in \code{feature_dt}}

\item{custom_range}{character vector that is a subset of levels in
\code{feature_col}}
}
\description{
Given a categorical column in a data.table or a custom range, generate
uniformly distributed points along its range.
}
\examples{
\dontrun{
dt <- data.table(a = 1:3, b = 4:6, c = LETTERS[1:3])
generate_grid_range_numeric_(dt, "c")
}
}

