% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependency_helper.R
\name{generate_grid_range_numeric_}
\alias{generate_grid_range_numeric_}
\title{Internal function for use by calculate_partial_dependency}
\usage{
generate_grid_range_numeric_(feature_dt, feature_col, num_grid = 100,
  custom_range = NULL)
}
\arguments{
\item{feature_dt}{data.table containing features used in predictive model}

\item{feature_col}{character. name of a column in \code{feature_dt}}

\item{num_grid}{number of points to distribute along range of
\code{feature_col} or \code{custom_range}}

\item{custom_range}{2-element numeric vector to be used instead of range of
\code{feature_col}}
}
\description{
Given a numeric column in a data.table or a custom range, generate
uniformly distributed points along its range.
}
\examples{
\dontrun{
dt <- data.table(a = 1:3, b = 4:6)
generate_grid_range_numeric_(dt, "a", 3)
}
}

