% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependency_aggregate.R
\name{loop_calculate_partial_dependency}
\alias{loop_calculate_partial_dependency}
\title{Convenience function to calculate partial dependency over multiple features}
\usage{
loop_calculate_partial_dependency(feature_dt, model_list,
  feature_cols = names(feature_dt), predict_fcn = predict,
  ensemble_fcn = median, ...)
}
\arguments{
\item{feature_dt}{data.table containing features used in predictive model}

\item{model_list}{named list of model objects. Each name will become a column
containing predictions from that model.}

\item{feature_cols}{character vector of column names in \code{feature_dt}
on which to calculate variable importance.
Defaults to all columns in \code{feature_dt}}

\item{predict_fcn}{function that accepts a model as its first argument
and \code{newdata} as one of its named arguments}

\item{ensemble_fcn}{function that combines a vector of predictions into
a single ensemble. Default is \code{median}}

\item{...}{additional arguments to pass to \code{\link{calculate_partial_dependency}}}
}
\value{
a list of \code{data.tables} returned by \code{\link{calculate_partial_dependency}}
}
\description{
Loops through feature columns in a \code{data.table} and
calculates partial dependency for each
}
\examples{
\dontrun{
dt <- data.table(a = 1:3, b = 2:4, c = c(8, 11, 14))
m <- lm(c ~ a + b - 1, dt)
gm <- glm(c ~ a + b - 1, data = dt)
loop_calculate_partial_dependency(dt, feature_cols = c("a", "b"),
                                  model_list = list(lm1 = m), num_grid = 6)
}
}

