% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependency_aggregate.R
\name{loop_calculate_pd_vimp}
\alias{loop_calculate_pd_vimp}
\title{Convenience function to calculate variable importance over multiple features}
\usage{
loop_calculate_pd_vimp(pd_list, vimp_colname = "ensemble")
}
\arguments{
\item{pd_list}{output from \code{\link{loop_calculate_partial_dependency}}}

\item{vimp_colname}{name of model (taken from the column names in \code{pd})
for which to calculate variable importance}
}
\value{
a list the length of \code{pd_list}, containing the variable importance
of each feature
}
\description{
Variable importance is calculated as the difference between the min and max
of the partial dependency output for a given feature
}
\examples{
\dontrun{
# Example output from loop_calculate_partial_dependency
pd <- list(a = data.table(feature = c("a", "a", "a"),
                          feature_val = c(1, 3.5, 6),
                          model1 = c(-2.5, 0, 2.5),
                          model2 = c(0, 0, 0),
                          ensemble = c(-2.5, -0.75, 0)),
           b = data.table(feature = c("b", "b", "b"),
                          feature_val = c(2, 3, 4),
                          model1 = c(0, 0, 0),
                          model2 = c(0, 0, 0),
                          ensemble = c(1, 2, 3)))
loop_calculate_pd_vimp(pd, vimp_colname = "ensemble")
}
}

