% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependency_aggregate.R
\name{loop_plot_fcn}
\alias{loop_plot_fcn}
\title{Generate partial dependency plots for multiple features}
\usage{
loop_plot_fcn(pd_list, plot_fcn = plot_partial_dependency)
}
\arguments{
\item{pd_list}{output from \code{\link{loop_calculate_partial_dependency}}}

\item{plot_fcn}{a function that accepts the output from
\code{\link{calculate_partial_dependency}} and returns a ggplot object.
This function will be looped over \code{pd_list}.}
}
\value{
a list of ggplot objects
}
\description{
Run \code{plot_fn} over output from
\code{loop_calculate_partial_dependency}
}
\examples{
\dontrun{
# Example output from loop_calculate_partial_dependency
pd <- list(a = data.table(feature = c("a", "a", "a"),
                          feature_val = c(1, 3.5, 6),
                          model1 = c(-2.5, 0, 2.5),
                          model2 = c(0, 0, 0),
                          ensemble = c(-2.5, -0.75, 0)),
           b = data.table(feature = c("a", "a", "a"),
                          feature_val = c(2, 3, 4),
                          model1 = c(0, 0, 0),
                          model2 = c(0, 0, 0),
                          ensemble = c(1, 2, 3)))
loop_plot_fcn(pd)
}
}

