% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variable_importance.R
\name{oob_importance}
\alias{oob_importance}
\title{Calculate the importance of all predictor variables in a training set}
\usage{
oob_importance(vars = names(x), model, x, y, inds, tuneparams, f,
  seed = sample(.Random.seed, 1), ...)
}
\arguments{
\item{vars}{character vector specifying variables for which to determine importance
Defaults to all predictor variables in \code{x}.}

\item{model}{character string defining method to pass to caret}

\item{x}{data.table containing predictor variables}

\item{y}{vector containing target variable}

\item{inds}{a list of integer vectors corresponding to the rows
used for each training iteration}

\item{tuneparams}{a data.frame containing hyperparameter values for caret.
Should only contain one value for each hyperparamter}

\item{f}{loss function to evaluate accuracy of model}

\item{seed}{set random seed for train method}

\item{...}{additional arguments to pass to caret train}
}
\description{
Deterimne variable importance of predictor variables by observing
the change in error when removing a single variable at a time
from training data
}
