% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependency_individual.R
\name{plot_partial_dependency}
\alias{plot_partial_dependency}
\title{Generate a partial dependency plot}
\usage{
plot_partial_dependency(pd)
}
\arguments{
\item{pd}{output from \code{\link{calculate_partial_dependency}}}
}
\description{
Take the output from \code{calculate_partial_dependency} and plot the
partial dependencies for every model.
This function returns the plot object so it can be further modified.
}
\examples{
\dontrun{
# Example output from calculate_partial_dependency
pd <- data.table(feature = rep("a", 9),
                 feature_val = rep(c(1, 3.5, 6), 3),
                 model = rep(c("model1",
                               "model2",
                               "ensemble"),
                             each = 3),
                 prediction = c(c(-2.5, 0, 2.5),
                                c(0, 0, 0),
                                c(-2.5, -0.75, 0)))
plot_partial_dependency(pd)
}
}

