% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependency_aggregate.R
\name{run_partial_dependency}
\alias{run_partial_dependency}
\title{Convenience function for calculating partial dependency and variable importance
over multiple features}
\usage{
run_partial_dependency(feature_dt, model_list,
  feature_cols = names(feature_dt), predict_fcn = predict,
  ensemble_colname = "ensemble", ensemble_fcn = median,
  ensemble_models = names(model_list), num_grid = 10, custom_range = NULL,
  plot_fcn = plot_partial_dependency, vimp_colname = "ensemble",
  plot = TRUE, facet = TRUE, ncol = NULL)
}
\arguments{
\item{feature_dt}{data.table containing features used in predictive model}

\item{model_list}{named list of model objects. Each name will become a column
containing predictions from that model.}

\item{feature_cols}{character vector of column names in \code{feature_dt}
on which to calculate variable importance.
Defaults to all columns in \code{feature_dt}}

\item{predict_fcn}{function that accepts a model as its first argument
and \code{newdata} as one of its named arguments}

\item{ensemble_colname}{character. Name of the column containing ensemble predictions}

\item{ensemble_fcn}{function that combines a vector of predictions into
a single ensemble. Default is \code{median}}

\item{ensemble_models}{character vector of names from model_list. These models
will be combined by ensemble_fcn to form the ensemble}

\item{num_grid}{number of points to distribute along range of
\code{feature_col} or \code{custom_range}}

\item{custom_range}{should only be used if \code{feature_cols} is a 1-element vector
Defines a custom range to calculate partial dependency over. This can
be a 2-element numerical vector or a character vector, depending on the
type of \code{feature_cols[1]}}

\item{plot_fcn}{a function that accepts the output from
\code{\link{calculate_partial_dependency}} and returns a ggplot object.}

\item{vimp_colname}{name of model (taken from from \code{model_list} or
\code{ensemble_colname}) for which to calculate variable importance}

\item{plot}{TRUE/FALSE. Should the partial dependencies be plotted? Defaults to TRUE}

\item{facet}{TRUE/FALSE. If \code{plot = TRUE}, should the graphs be combined
into one plot? Defaults to TRUE}

\item{ncol}{if \code{facet = TRUE}, number of columns in the facetted plot}
}
\value{
Output is a \code{data.table} with one column for every model in
\code{model_list}, an ensemble column, feature name and feature value columns,
and the variable importance column
}
\description{
Given a list of models, get their average prediction over a range of
values for each feature in \code{features_cols}
}
\examples{
\dontrun{
dt <- data.table(a = 1:3, b = 2:4, c = c(8, 11, 14))
m <- lm(c ~ a + b - 1, dt)
gm <- glm(c ~ a + b - 1, data = dt)
run_partial_dependency(feature_dt = dt[, list(a, b)],
                       model_list = list(lm1 = m, gm1 = gm))
}
}

