% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variable_importance.R
\name{varImpBag}
\alias{varImpBag}
\title{Internal function for use by oob_importance}
\usage{
varImpBag(var, model, x, y, inds, tuneparams, f, loss_list, seed, ...)
}
\arguments{
\item{var}{character. variable to remove from training data
in order to obtain importance of that variable.}

\item{model}{character string defining method to pass to caret}

\item{x}{data.table containing predictor variables}

\item{y}{vector containing target variable}

\item{inds}{a list of integer vectors corresponding to the rows
used for each training iteration}

\item{tuneparams}{a data.frame containing hyperparameter values for caret.
Should only contain one value for each hyperparamter.
If method does not require hyperparamter settings, set to NULL}

\item{f}{loss function to evaluate accuracy of model}

\item{loss_list}{list containing basline error of each sampling iteration}

\item{seed}{set random seed for train method}

\item{...}{additional arguments to pass to caret train}
}
\description{
Calculate the change in error from baseline model
when removing a variable from training data.
}
