% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variable_importance.R
\name{calculate_marginal_vimp}
\alias{calculate_marginal_vimp}
\title{Calculate the marginal importance of variables in a predictor matrix}
\usage{
calculate_marginal_vimp(x, y, method, loss_metric, resampling_indices, tuneGrid,
  vars = names(x), seed = sample(.Random.seed, 1), allow_parallel = FALSE,
  ...)
}
\arguments{
\item{x}{data.table containing predictor variables}

\item{y}{vector containing target variable}

\item{method}{character string defining method to pass to caret}

\item{loss_metric}{character. Loss metric to evaluate accuracy of model}

\item{resampling_indices}{a list of integer vectors corresponding to the row indices
used for each resampling iteration}

\item{tuneGrid}{a data.frame containing hyperparameter values for caret.
Should only contain one value for each hyperparameter. Set to NULL
if caret method does not have any hyperparameter values.}

\item{vars}{character vector specifying variables for which to determine marginal importance
Defaults to all predictor variables in \code{x}.}

\item{seed}{set random seed for train method}

\item{allow_parallel}{boolean for parallel execution. If set to TRUE, user must specify parallel backend
in their R session to take advantage of multiple cores. Defaults to FALSE.}

\item{...}{additional arguments to pass to caret train}
}
\description{
A caret model is trained on training data according to resampling indices
specified by the user, and error is calculated on out of sample data.
Variable importance is determined by calculating the change in out of sample model performance
when a variable is removed relative to baseline out of sample performance when all
variables are included.
}
