% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_partial_dependency.R
\name{calculate_partial_dependency}
\alias{calculate_partial_dependency}
\title{Calculate partial dependency for list of models and optionally plot it}
\usage{
calculate_partial_dependency(feature_dt, feature_col, model_list,
  num_grid = 10, custom_range = NULL, predict_fcn = predict,
  ensemble_colname = "ensemble", ensemble_fcn = median,
  ensemble_models = names(model_list), plot = TRUE)
}
\arguments{
\item{feature_dt}{data.table containing features used in predictive model}

\item{feature_col}{character. name of a column in \code{feature_dt}}

\item{model_list}{named list of model objects. Each name will become a column
containing predictions from that model.}

\item{num_grid}{number of points to distribute along range of
\code{feature_col} or \code{custom_range}}

\item{custom_range}{2-element numeric vector to be used instead of range of
\code{feature_col}}

\item{predict_fcn}{function that accepts a model as its first argument
and \code{newdata} as one of its named arguments}

\item{ensemble_colname}{character. Name of the column containing ensemble predictions}

\item{ensemble_fcn}{function that combines a vector of predictions into
a single ensemble. Default is median}

\item{ensemble_models}{character vector of names from model_list. These models
will be combined by ensemble_fcn to form the ensemble}

\item{plot}{TRUE/FALSE Should the partial dependencies be plotted? Defaults to TRUE}
}
\description{
Given a list of models, get their average prediction over a range of
values from a specified feature
}
\examples{
\dontrun{
dt <- data.table(a = 1:3, b = 2:4, c = c(8, 11, 14))
m <- lm(c ~ a + b - 1, dt)
gm <- glm(c ~ a + b - 1, data = dt)
calculate_partial_dependency(dt, "a", list(lm1 = m),
                             num_grid = 6)
calculate_partial_dependency(dt, "a", list(lm1 = m, glm1 = gm),
                             num_grid = 6, ensemble_fcn = sum)
calculate_partial_dependency(dt, "a", list(lm1 = m),
                             num_grid = 6, custom_range = c(1,6))
}
}

