% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_partial_dependency.R
\name{calculate_pd_vimp}
\alias{calculate_pd_vimp}
\title{Calculate variable importance based on partial dependency for list of models}
\usage{
calculate_pd_vimp(feature_dt, model_list, vimp_colname = "ensemble",
  feature_cols = names(feature_dt), allow_parallel = FALSE, ...)
}
\arguments{
\item{feature_dt}{data.table containing features used in predictive model}

\item{model_list}{named list of model objects. Each name will become a column
containing predictions from that model.}

\item{vimp_colname}{name of model (taken from from \code{model_list} or
\code{ensemble_colname}) for which to calculate variable importance}

\item{feature_cols}{character vector of column names in \code{feature_dt}
on which to calculate variable importance.
Defaults to all columns in \code{feature_dt}}

\item{allow_parallel}{boolean for parallel execution.
If set to TRUE, user must specify parallel backend in their R session
to take advantage of multiple cores. Defaults to FALSE.}

\item{...}{additional arguments to pass to  \code{\link{calculate_partial_dependency}}}
}
\description{
Given a list of models, get their average prediction over a range of
values for each feature in \code{features_cols}
}
\examples{
\dontrun{
TODO: fill this out
}
}

