% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependency_individual.R
\name{calculate_pd_vimp_normed}
\alias{calculate_pd_vimp_normed}
\title{Calculate normed variable importance based on partial dependency,
taking into account model variability}
\usage{
calculate_pd_vimp_normed(pd, ensemble_colname = "ensemble", epsilon = 1e-07)
}
\arguments{
\item{pd}{output from \code{\link{calculate_partial_dependency}}}

\item{ensemble_colname}{name of ensemble column specified in \code{\link{calculate_partial_dependency}}.}

\item{epsilon}{small value to add to the minimum standard deviation before dividing to prevent Inf return value.
Defaults to 1e-07.}
}
\description{
Normed variable importance is calculated as the difference between the min and max,
divided by the minimum standard deviation of the individual model predictions at any value cutpoint
Most effective when training many models on different subsamples of training data.
}
\examples{
\dontrun{
# Example output from calculate_partial_dependency
pd <- data.table(feature = rep("a", 9),
                 feature_val = rep(c(1, 3.5, 6), 3),
                 model = rep(c("model1",
                               "model2",
                               "ensemble"),
                             each = 3),
                 prediction = c(c(-2.5, 0, 2.5),
                                c(0, 0, 0),
                                c(-2.5, -0.75, 0)))
calculate_pd_vimp_normed(pd, ensemble_colname = "ensemble")
}
}

