% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependency_aggregate.R
\name{facet_plot_fcn}
\alias{facet_plot_fcn}
\title{Multiple partial dependency plots on one page}
\usage{
facet_plot_fcn(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot2 objects}

\item{plotlist}{a list of ggplot2 objects}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\description{
It can be helpful to view multiple partial dependency plots at once
to get a sense of variable importance, relationships between variables, etc.
}
\details{
If the layout is something like \code{matrix(c(1,2,3,3), nrow=2, byrow=TRUE)},
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}

