% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependency_aggregate.R
\name{loop_plot_fcn}
\alias{loop_plot_fcn}
\title{Generate partial dependency plots for multiple features}
\usage{
loop_plot_fcn(pd_list, plot_fcn = plot_partial_dependency)
}
\arguments{
\item{pd_list}{output from \code{\link{loop_calculate_partial_dependency}}}

\item{plot_fcn}{a function that accepts the output from
\code{\link{calculate_partial_dependency}} and returns a ggplot object.
This function will be looped over \code{pd_list}.}
}
\value{
a list of ggplot objects
}
\description{
Run \code{plot_fn} over output from
\code{loop_calculate_partial_dependency}
}
\examples{
\dontrun{
# Example output from loop_calculate_partial_dependency
pd <- list(a = data.table(feature = rep("a", 9),
                          feature_val = rep(c(1, 3.5, 6), 3),
                          model = rep(c("model1",
                                        "model2",
                                        "ensemble"),
                                      each = 3),
                          prediction = c(c(-2.5, 0, 2.5),
                                       c(0, 0, 0),
                                       c(-2.5, -0.75, 0))),
           b = data.table(feature = rep("b", 9),
                          feature_val = rep(c(2, 3, 4), 3),
                          model = rep(c("model1",
                                        "model2",
                                        "ensemble"),
                                      each = 3),
                          prediction = c(c(0, 0, 0),
                                       c(0, 0, 0),
                                       c(1, 2, 3))))
loop_plot_fcn(pd)
}
}

